Move the GenericCreep definition to a separate file
