+++ src/main/kotlin/screeps/api/Traits.kt
@@ -46,47 +46,3 @@ external interface Store {
     val store: StoreDefinition
     val storeCapacity: Int
 }
-
-external interface GenericCreep : RoomObjectNotNull, Owned, Attackable, Identifiable {
-    val carry: StoreDefinition
-    val carryCapacity: Int
-    val memory: MemoryMarker
-    val name: String
-    val saying: String
-    val ticksToLive: Int?
-
-    fun drop(resourceType: ResourceConstant, amount: Int = definedExternally): ScreepsReturnCode
-    fun move(direction: DirectionConstant): ScreepsReturnCode
-    fun moveByPath(path: Array<Room.PathStep>): ScreepsReturnCode
-    fun moveByPath(serializedPath: String): ScreepsReturnCode
-    fun moveTo(target: NavigationTarget, opts: MoveToOptions = definedExternally): ScreepsReturnCode
-    fun moveTo(x: Int, y: Int, opts: MoveToOptions = definedExternally): ScreepsReturnCode
-    fun notifyWhenAttacked(enable: Boolean): ScreepsReturnCode
-    fun pickup(target: Resource): ScreepsReturnCode
-    fun say(message: String, toPublic: Boolean = definedExternally): ScreepsReturnCode
-    fun suicide(): ScreepsReturnCode
-    fun transfer(target: Structure, resourceType: ResourceConstant, amount: Int = definedExternally): ScreepsReturnCode
-    fun transfer(target: GenericCreep, resourceType: ResourceConstant, amount: Int = definedExternally): ScreepsReturnCode
-
-    fun withdraw(
-            target: Structure,
-            resourceType: ResourceConstant,
-            amount: Int = definedExternally
-    ): ScreepsReturnCode
-
-    fun withdraw(
-            target: Tombstone,
-            resourceType: ResourceConstant,
-            amount: Int = definedExternally
-    ): ScreepsReturnCode
-
-
-
-}
-
-external interface MoveToOptions : FindPathOptions {
-    val reusePath: Int? get() = definedExternally
-    val serializeMemory: Boolean? get() = definedExternally
-    val noPathFinding: Boolean? get() = definedExternally
-    val visualizePathStyle: RoomVisual.Style? get() = definedExternally
-}
